+++ tests/views/test_comments_order_view.py
@@ -22,7 +22,6 @@ class TestOrderComments(BaseTestView):
     model = models.Comment
     UPDATE_SUCCESS_MESSAGE = ''
     NOT_FOUND_MESSAGE = ''
-    ignore_validation_fields = ['entity', 'state', 'state_history']
     payload_position = 10
     update_map = {
         'content': 'new message content',